Refactor date validation and emit logic
